+++ app/src/main/java/com/example/flavoryapp/FavoriteFragment.kt
@@ -1,0 +1,20 @@
+package com.example.flavoryapp
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.flavoryapp.databinding.FragmentFavoriteBinding
+
+class FavoriteFragment : Fragment() {
+    private lateinit var mBinding: FragmentFavoriteBinding
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        mBinding = FragmentFavoriteBinding.inflate(inflater)
+        return mBinding.root
+    }
+}
