+++ app/src/main/java/com/example/flavoryapp/HomeFragment.kt
@@ -1,0 +1,20 @@
+package com.example.flavoryapp
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.flavoryapp.databinding.FragmentHomeBinding
+
+class HomeFragment : Fragment() {
+    private lateinit var binding: FragmentHomeBinding
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        binding = FragmentHomeBinding.inflate(inflater)
+        return binding.root
+    }
+}
