+++ app/src/main/java/com/example/flavoryapp/ProfileFragment.kt
@@ -1,4 +1,4 @@
-package com.example.flavoryapp.uis.view
+package com.example.flavoryapp
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,7 +8,6 @@ import android.view.ViewGroup
 import com.example.flavoryapp.databinding.FragmentProfileBinding
 
 class ProfileFragment : Fragment() {
-
     private lateinit var mBinding: FragmentProfileBinding
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -18,5 +17,4 @@ class ProfileFragment : Fragment() {
         mBinding = FragmentProfileBinding.inflate(inflater, container, false)
         return mBinding.root
     }
-
 }
