--- app/src/main/java/com/example/flavoryapp/data/api/RetrofitBuilder.kt
@@ -1,38 +1,0 @@
-package com.example.flavoryapp.data.api
-
-import okhttp3.HttpUrl
-import okhttp3.Interceptor
-import okhttp3.OkHttpClient
-import okhttp3.Request
-import okhttp3.Response
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object RetrofitBuilder {
-
-    private const val BASE_URL = "https://api.spoonacular.com/"
-//    private const val API_KEY = "a92ba302332043249daf5d02b0620791"
-    private const val API_KEY = "3324cb9d269e41099407750992f1a9fe"
-
-    private fun getRetrofit() = Retrofit.Builder()
-        .baseUrl(BASE_URL)
-        .addConverterFactory(GsonConverterFactory.create())
-        .client(httpClient.build())
-        .build()
-
-//    Adds API KEY to every request
-   private val httpClient : OkHttpClient.Builder = OkHttpClient.Builder()
-       .addInterceptor(Interceptor { chain ->
-           val originalRequest : Request = chain.request()
-           val newRequest = originalRequest.newBuilder()
-               .addHeader("x-api-key", API_KEY)
-               .build()
-           chain.proceed(newRequest)
-       })
-       .addInterceptor((HttpLoggingInterceptor().apply {
-        level = HttpLoggingInterceptor.Level.BASIC // Set the logging level to BASIC
-    }))
-
-    val apiService : ApiService = getRetrofit().create(ApiService::class.java)
-}
