--- app/src/main/java/com/example/flavoryapp/data/model/AnalyzedInstructions.kt
@@ -1,8 +1,0 @@
-package com.example.flavoryapp.data.model
-
-import com.example.flavoryapp.uis.view.IngredientInstructionInterface
-
-data class AnalyzedInstructions (
-    var name  : String?          = null,
-    var steps : ArrayList<Steps>? = arrayListOf()
-): IngredientInstructionInterface
