--- app/src/main/java/com/example/flavoryapp/data/model/Nutrients.kt
@@ -1,8 +1,0 @@
-package com.example.flavoryapp.data.model
-
-data class Nutrients (
-    var name                : String? = null,
-    var amount              : Double?    = null,
-    var unit                : String? = null,
-    var percentOfDailyNeeds : Double?    = null
-)
