--- app/src/main/java/com/example/flavoryapp/data/model/Nutrition.kt
@@ -1,5 +1,0 @@
-package com.example.flavoryapp.data.model
-
-data class Nutrition (
-    var nutrients        : ArrayList<Nutrients>?   = arrayListOf()
-)
