--- app/src/main/java/com/example/flavoryapp/data/model/RecipeResponseData.kt
@@ -1,5 +1,0 @@
-package com.example.flavoryapp.data.model
-
-data class RecipeResponseData (
-    var recipes      : ArrayList<RecipeData> = arrayListOf(),
-)
