--- app/src/main/java/com/example/flavoryapp/data/model/SearchedRecipeData.kt
@@ -1,9 +1,0 @@
-package com.example.flavoryapp.data.model
-
-data class SearchedRecipeData(
-    var offset: Int? = null,
-    var number: Int? = null,
-    var results: ArrayList<SearchedRecipe>? = arrayListOf(),
-    var totalResults: Int? = null,
-    var recipeDataToStore : RecipeData
-)
