--- app/src/main/java/com/example/flavoryapp/data/model/Steps.kt
@@ -1,8 +1,0 @@
-package com.example.flavoryapp.data.model
-
-import com.example.flavoryapp.uis.view.IngredientInstructionInterface
-
-data class Steps (
-    var number      : Int?                   = null,
-    var step        : String?                = null
-): IngredientInstructionInterface
