--- app/src/main/java/com/example/flavoryapp/uis/intent/UnitIntent.kt
@@ -1,10 +1,0 @@
-package com.example.flavoryapp.uis.intent
-
-sealed class UnitIntent {
-    data class ConvertUnit(
-        val toConvertFrom: String,
-        val toConvertInto: String,
-        val amount: String
-    ) :
-        UnitIntent()
-}
