--- app/src/main/java/com/example/flavoryapp/uis/view/IngredientInstructionInterface.kt
@@ -1,4 +1,0 @@
-package com.example.flavoryapp.uis.view
-
-interface IngredientInstructionInterface {
-}
