--- app/src/main/java/com/example/flavoryapp/uis/viewstate/UnitState.kt
@@ -1,8 +1,0 @@
-package com.example.flavoryapp.uis.viewstate
-
-sealed class UnitState {
-    object Inactive : UnitState()
-    object Loading : UnitState()
-    data class ResponseData(val convertedAmt: String) : UnitState()
-    data class Error(val error: String?) : UnitState()
-}
